# docs/swagger.yaml
info:
  contact: {}
paths:
  /api/v1/article:
    post:
      parameters:
      - description: 标签id
        in: query
        name: tag_id
        required: true
        type: integer
      - description: 标题
        in: query
        name: title
        required: true
        type: string
      - description: 简述
        in: query
        name: desc
        required: true
        type: string
      - description: 内容
        in: query
        name: content
        required: true
        type: string
      - description: State:状态只允许0或1
        in: query
        name: state
        type: integer
      - description: 创建人
        in: query
        name: created_by
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"ok","data":{}}'
          schema:
            type: string
      summary: 新增文章
  /api/v1/article/{id}:
    delete:
      parameters:
      - description: 文章id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"ok","data":{}}'
          schema:
            type: string
      summary: 删除文章
    get:
      parameters:
      - description: ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"ok","data":{}}'
          schema:
            type: string
      summary: 获取单个文章
    put:
      parameters:
      - description: 文章id
        in: path
        name: id
        required: true
        type: integer
      - description: 标签id
        in: query
        name: tag_id
        required: true
        type: integer
      - description: 标题
        in: query
        name: title
        required: true
        type: string
      - description: 简述
        in: query
        name: desc
        required: true
        type: string
      - description: 内容
        in: query
        name: content
        required: true
        type: string
      - description: State:状态只允许0或1
        in: query
        name: state
        type: integer
      - description: 修改人
        in: query
        name: modified_by
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"ok","data":{}}'
          schema:
            type: string
      summary: 编辑文章
  /api/v1/articles:
    get:
      parameters:
      - description: State
        in: query
        name: state
        type: integer
      - description: 标签id
        in: query
        name: tag_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"ok","data":{}}'
          schema:
            type: string
      summary: 获取多个文章
  /api/v1/tag:
    post:
      parameters:
      - description: Name
        in: query
        name: name
        required: true
        type: string
      - description: State
        in: query
        name: state
        type: integer
      - description: CreatedBy
        in: query
        name: created_by
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"ok","data":{}}'
          schema:
            type: string
      summary: 新增文章标签
  /api/v1/tag/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"ok","data":{}}'
          schema:
            type: string
      summary: 删除文章标签
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Name
        in: query
        name: name
        required: true
        type: string
      - description: State
        in: query
        name: state
        type: integer
      - description: ModifiedBy
        in: query
        name: modified_by
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"ok","data":{}}'
          schema:
            type: string
      summary: 修改文章标签
  /api/v1/tags:
    get:
      parameters:
      - description: Name
        in: query
        name: name
        required: true
        type: string
      - description: State
        in: query
        name: state
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"ok","data":{}}'
          schema:
            type: string
      summary: 获取文章标签列表
swagger: "2.0"